Add an internal.Main function.
This will be used to intercept and wrap incoming HTTP requests when the
app uses appengine packages. go-app-builder will be modified to
recognize when an app includes github.com/golang/appengine/internal, and
generate a synthetic main that calls this.

Right now this only operates the same way as go-app-builder's standalone
main, but permits us to go modify go-app-builder now and roll that
change out independent of the more complex changes here.

Change-Id: I7dc40f165f8947d54dedc81a890fe248fc1b341a