Fix race condition in TestAPICallRPCFailure.
For the test that checked behaviour of API calls that time out,
there was no synchronisation between the HTTP handler starting
and the httptest.Server teardown at the end of the test (I/O only
went in one direction). Replacing the unnecessary f.die chan with
an f.hang chan that is explicitly strobed in the HTTP handler fixes
the race.

Change-Id: I3d3c2e92a3561a9795a8d4ebb1a9ba4df8a8d6d6