search: ignore struct fields with "-" struct tags
This change brings search in line with datastore's behaviour.

Also, fixes a bug where loading into a struct with field mismatch
would not return an error if there was no document metadata
(this case could only happen in unit tests, since the metadata is
never nil when called from the public API).

Change-Id: Ia606f94e38bf66b38c27a44d0f1fdf5e789d9545