Normalize namespace and transaction application.
This causes Call to apply Namespace at the very beginning of the
Call chain.

This prevents an issue where some appserver APIs treat NameSpace=nil
and NameSpace="" distinctly (with the latter being an error). Now if
the namespace is unset or set to "" in the SDK, no namespace will be
applied to the outgoing message.

This also implements a very basic ContextForTesting for classic
testing.

Change-Id: I0c5c2917d958920ab015b8cf7f6b34a2a1b8d2a0