cmd/aedeploy: rewrite import scanning logic
The new implementation is largely copied from github.com/broady/deplist.

It uses the go/build package to find imports instead of scanning GOPATH
directly. This enables the use of vendored dependencies as well as
using a more robust detection of imports from the standard library.

Some more work needs to be done (more thought needed) on handling
dependencies that are vendored from multiple locations.

For example:
  a/
    vendor/
      c/
    aeapp/ // package main
      app.yaml
      main.go // import "c" and import "b"
  b/
    b.go // import "c"
  vendor/
    c/

This structure has a dependency on "a/vendor/c" and a transitive
dependency on "vendor/c" via "b".

Fixes #29
Fixes #30

Change-Id: I44034f6559db6cddd1f28361ce438ec23fec35de