internal: make log calls panic with a useful error message
Prior to this change, a program would panic with no useful message:
panic: runtime error: invalid memory address or nil pointer dereference

Use the same "not an App Engine context" message that is returned as an
error from other API calls. The log package is special because the
logs are flushed asynchronously, and thus the API call happens later.

Change-Id: I896d8463d2a0d5da7a4fc94fd0dab7b6fcfa8ad7