internal: give more useful panics when (mis)using context.Background (#68)
Instead of waiting for a nil return from fromContext to panic, I've
updated all callers to proactively return/panic with a common error
type.  Panic is appropriate here, as it is a failed precondition of
values passed into the functions.  It's not introducing any new panics,
just ones with more useful messages.

Addresses GoogleCloudPlatform/google-cloud-go#523