log: clarify package docs
The package docs can be interpreted as: "this package is only for
querying the logs". Unless reader looks at actual func names, it's not
obvious they can actually write the logs as well.

It would be nice to mention this package in any of the tutorials as
well, since stderr output just disappears.