delay: permit Context from either stdlib or golang.org/x/net (#95)
Go 1.7 introduced the Context type into the standard library, but it is
assignable to the golang.org/x/net variant.  The appengine libraries are
written in terms of golang.org/x/net, which can interoperate with the
standard library type most of time, but because delay uses reflection,
it depends directly on the exact type.

The test in this change only failed on Go 1.7 and Go 1.8 but not Go
1.9, since after Go 1.9 the golang.org/x/net type became a type alias
for the standard library type.