datastore: add AllocateIDRange
Allows a range of IDs to be allocated, as is possible with the Python
and Java appengine APIs. Note that because of the way the datastore
backend is implemented, a request to allocate ID range [x, y] is
actually a request to allocate range [1, y]. This is consistent with
the behavior in Python and Java.

Test: Because we don't have a mock datastore, no tests are added to
      key_test.go (AllocateIDs is also not covered in key_test.go).
      Instead, I made test apps in Go (using these changes) and Python
      (using stock google.appengine.ext.db) and exercised each API's
      ID range allocation implementation with the dev_appserver.py
      datastore and the prod datastore. The Go implementation matches
      Python's behavior.

Fixes #97