Add functions that check which environment the app is running in
This is motivated by the fact that second-generation runtimes
(>= Go 1.11) do not set the appengine build tag, so the build tag alone
is not sufficient to determine whether an app is running in the standard
environment. We can check for a second-gen runtime with $GAE_ENV.
The new functions centralize this logic so other packages don't have to
worry about the build tags and env vars.

Added:
- IsStandard reports whether the App Engine app is running in the
  standard environment. This includes both the first generation
  runtimes (<= Go 1.9) and the second generation runtimes (>= Go 1.11).
- IsFlex reports whether the App Engine app is running in the flexible
  environment.
- IsAppEngine reports whether the App Engine app is running on App
  Engine, in either the standard or flexible environment.

Test: Deployed an app to go1/go1.8, go1/go1.9, go111, and flex. The app
      prints out the return val of each of the new functions. Verified
      correctness.