aetest: Redirect stderr to be visible to the user
Previously, all stderr after the admin/API server lines was lost as the
StderrPipe continued to consume it without redirecting once the scanner
went out of scope. It also had the side effect of hanging a program in
which the devserver emitted too much data to stderr, since the pipe
would block once it reached cap.

This replaces the TeeReader (which would go out of scope before stderr
was done) by printing the output while processing and copying it once
processing finishes.