[internal] Add LOG_TO_LOGSERVICE env var to disable logservice (#229)
Set env var LOG_TO_LOGSERVICE=0 to prevent sending logs to the App Engine logservice.

The longer term goal is to remove all of the code behind this flag and replace it with structured json logs. The gen2 runtimes watch /var/log for structured logfiles and forwards their records to stackdriver.