fix: regenerate all protos under new namespace
This should be wire compat as field numbers did not change and
package is not apart of the serialization. I appened `.v2` to each
proto package so that these protos can register distinctly from the
protos that were copied from the v1 library. This removes runtime
errors due to double registration of the same messages. Note: I did
have to specially regenerate taskqueue with a special built version
of protoc-gen-go in order to generate some proto1 legacy message
type that is referenced.

Fixes: #311
Fixes: https://github.com/googleapis/google-cloud-go/issues/7760